--- lib/core/constants.dart
@@ -1,29 +1,0 @@
-import 'package:flutter/material.dart';
-
-// Font Colors
-const fontColorBlack = Color.fromRGBO(43, 45, 51, 1);
-const fontColorTitle = Color.fromRGBO(255, 255, 255, 1);
-const fontColorSubtitle = Color.fromRGBO(237, 240, 246, 1);
-const fontColorError = Color.fromRGBO(200, 20, 20, 1);
-
-// Font Weights
-const fontWeightRegular = FontWeight.w400;
-const fontWeightMedium = FontWeight.w500;
-const fontWeightBold = FontWeight.w700;
-
-// Weather Background Colors
-const backgroundShiny = Color.fromRGBO(92, 162, 216, 1);
-const backgroundCloudy = Color.fromRGBO(125, 169, 201, 1);
-const backgroundNightCloudy = Color.fromRGBO(26, 54, 125, 1);
-const backgroundRainy = Color.fromRGBO(18, 48, 91, 1);
-const backgroundSnowy = Color.fromRGBO(109, 132, 162, 1);
-const backgroundThunder = Color.fromRGBO(5, 37, 120, 1);
-const backgroundElseWeather = Color.fromRGBO(13, 49, 141, 1);
-const backgroundEnd = Color.fromRGBO(83, 118, 172, 1);
-
-// Tasks' Cards' Colors
-const backgroundColor = Color.fromRGBO(255, 255, 255, 1);
-const backgroundBox = Color.fromRGBO(255, 255, 255, 0.15);
-
-// Values
-const mainScreenTaskListLimit = 4;
