+++ lib/data/models/weather/weather_model.dart
