+++ lib/pages/main/bloc/main_page_event.dart
@@ -1,0 +1,11 @@
+part of 'main_page_bloc.dart';
+
+sealed class MainPageEvent {
+  MainPageEvent();
+}
+
+final class MainPageInitialEvent extends MainPageEvent {}
+
+final class MainPageOnChangeEvent extends MainPageEvent {
+  MainPageOnChangeEvent();
+}
